+++ src/configV2/configV2.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { ConfigV2Service } from './configV2.service';
+
+@Module({
+  imports: [],
+  controllers: [],
+  providers: [ConfigV2Service],
+  exports: [ConfigV2Service]
+})
+export class ConfigV2Module {}