+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-### 1.10.2
+### 1.10.3
 
 #### Features